pzdc kak vi eto shavali)?)X)X)X)X)))

+++ Tyuiu.KimmelDS.Sprint6.Task0.V13.Lib/DataService.cs
@@ -4,6 +4,6 @@ namespace Tyuiu.KimmelDS.Sprint6.Task0.V13.Lib
 {
     public class DataService : ISprint6Task0V13
     {
-        public double Calculate(int x) => (x * x * x) / (x * x - 1);
+        public double Calculate(int x) => (double)(x * x * x) / (double)(x * x - 1);
     }
 }
